Make the embed code a trait.
